Add comment systm spec

+++ db/migrate/20170619064321_add_comments_count_to_posts.rb
@@ -1,5 +1,5 @@
 class AddCommentsCountToPosts < ActiveRecord::Migration[5.1]
   def change
-    add_column :posts, :comments_count, :integer, default: 0
+    add_column :posts, :comments_count, :bigint, default: 0
   end
 end
